third_party/WebKit/LayoutTests/fast/filesystem/cross-filesystem-op.html

<html>
<body>
<script src="resources/fs-test-util.js"></script>
<div>This tests basic cross-filesystem operations.</div>

<div id="console"></div>

<script>
var methods = [ 'copyTo', 'moveTo' ];
var sourceTestDirectory, destTestDirectory;

var sourceDirectoryPath = '/resources';
var destDirectoryPath = '/test';

// Files to create.
var testFiles = [
    { path:'apple.gif', size: 1476 },
    { path:'mozilla.gif', size: 2593 },
    { path:'drop-file-svg.svg', size: 109 },
    { path:'copy-backslash-euc.html', size: 478 },
    { path:'test_directory', directory:true },
    { path:'test_directory/test.txt', size: 13 }
];

// TODO(jsbell): Test moveTo.
var tests = [
    testCopyFile,
    testCopyDirectory,
];

function log(text)
{
    var console = document.getElementById('console');
    console.appendChild(document.createTextNode(text));
    console.appendChild(document.createElement('br'));
}

function test(expect, actual)
{
    log((expect == actual ? 'PASS' : 'FAIL') + ': "' + expect + '" == "' + actual + '"');
}


function errorCallback(msg)
{
    return function(e) {
        console.log(e);
        log('ERROR:' + msg + ': ' + e.name);
        if (window.testRunner)
            testRunner.notifyDone();
    };
}

function setupSourceFileSystem(successCallback) {
    var bytes_needed = 1e5;
    navigator.webkitPersistentStorage.requestQuota(bytes_needed, function(granted) {
        webkitRequestFileSystem(window.PERSISTENT, bytes_needed, function (fs) {
            removeAllInDirectory(fs.root, function() {
                fs.root.getDirectory(
                    sourceDirectoryPath,
                    {create:true},
                    function (entry) {
                        sourceTestDirectory = entry;
                        successCallback();
                    },
                  errorCallback(`src: createDirectory(${sourceDirectoryPath})`));
              }, errorCallback('src: removeAllInDirectory'));
        }, errorCallback('src: requestFileSystem for PERSISTENT'));
    }, errorCallback('src: webkitPersistenStorage.requestQuota'));
}

function setupDestFileSystem(successCallback) {
    webkitRequestFileSystem(window.TEMPORARY, 1024, function (fs) {
        removeAllInDirectory(fs.root, function() {
            fs.root.getDirectory(
                destDirectoryPath,
                {create:true},
                function (entry) {
                    destTestDirectory = entry;
                    successCallback();
                },
              errorCallback(`dest: createDirectory(${destDirectoryPath})`));
          }, errorCallback('dest: removeAllInDirectory'));
    }, errorCallback('dest: requestFileSystem for TEMPORARY'));
}

function runTests(index, successCallback) {
    if (index >= tests.length) {
        successCallback();
        return;
    }
    var next = function() { runTests(index + 1, successCallback); };
    resetDirectory(destTestDirectory, function() { tests[index](next); });
}

function testCopyFile(successCallback) {
    testOnFile(0, 'copyTo', successCallback);
}

function testCopyDirectory(successCallback) {
    testOnDirectory(0, 'copyTo', successCallback);
}

function testOnFile(index, method, successCallback) {
    if (index >= testFiles.length) {
        successCallback();
        return;
    }
    var next = function() { testOnFile(index + 1, method, successCallback); };
    if (testFiles[index].directory) {
        next();
        return;
    }
    var file = testFiles[index];
    log('Testing ' + method + ' ' + file.path + ': ' + sourceTestDirectory.fullPath + ' => ' + destTestDirectory.fullPath);
    sourceTestDirectory.getFile(
        file.path, {create:false}, function(source) {
            source[method](destTestDirectory, '', function(dest) {
                verifyFile(source, dest, next);
            }, errorCallback('testOnFile: ' + method + ':' + file.path));
        }, errorCallback('testOnFile: getFile:' + file.path));
}

function testOnDirectory(index, method, successCallback) {
    if (index >= testFiles.length) {
        successCallback();
        return;
    }
    var next = function() { testOnDirectory(index + 1, method, successCallback); };
    if (!testFiles[index].directory) {
        next();
        return;
    }
    var file = testFiles[index];
    log('Testing ' + method + ' ' + file.path + ': ' + sourceTestDirectory.fullPath + ' => ' + destTestDirectory.fullPath);
    sourceTestDirectory.getDirectory(
        file.path, {create:false}, function(source) {
            source[method](destTestDirectory, '', function(dest) {
                verifyDirectory(source, dest, next);
            }, errorCallback('testOnDirectory: ' + method + ':' + file.path));
        }, errorCallback('testOnDirectory: getDirectory:' + file.path));
}

function verifyFile(source, dest, successCallback) {
    test(source.name, dest.name);
    source.getMetadata(function(sourceMetadata) {
        var expectedSize = sourceMetadata.size;
        dest.getMetadata(function(destMetadata) {
            test(expectedSize, destMetadata.size);
            successCallback();
        }, errorCallback('getMetadata:' + dest.fullPath));
    }, errorCallback('getMetadata:' + source.fullPath));
}

function verifyDirectory(source, dest, successCallback) {
    test(source.name, dest.name);
    getDirectoryEntries(source, function(sourceEntries) {
        getDirectoryEntries(dest, function(destEntries) {
            function byName(a, b) { return a.name < b.name ? -1 : b.name < a.name ? 1 : 0; }
            sourceEntries.sort(byName);
            destEntries.sort(byName);
            test(sourceEntries.length, destEntries.length);
            var verifyOne = function(index) {
                if (index >= sourceEntries.length) {
                    successCallback();
                    return;
                }
                verifyFile(sourceEntries[index], destEntries[index], function() {
                    verifyOne(index + 1);
                });
            };
            verifyOne(0);
        });
    });
}

function getDirectoryEntries(entry, successCallback) {
    var allEntries = [];
    var reader = entry.createReader();
    function nextChunk() {
        reader.readEntries(function(entries) {
            if (entries.length === 0) {
                successCallback(allEntries);
            } else {
                allEntries = allEntries.concat(entries);
                nextChunk();
            }
        }, errorCallback('readEntries: ' + entry.fullPath));
    }
    nextChunk();
}

function cleanupDirectory(directory, successCallback) {
    directory.removeRecursively(
        successCallback,
        errorCallback('removeRecursively:' + directory.fullPath));
}

function resetDirectory(directory, successCallback) {
    var fs = directory.filesystem;
    var path = directory.fullPath;
    cleanupDirectory(directory, function() {
        fs.root.getDirectory(path, {create:true},
                             function () {
                                log(fs.name + ': reset ' + path);
                                successCallback();
                             }, errorCallback('getDirectory(create):' + path));
    });
}

function setupFiles(root, index, successCallback)
{
    if (index >= testFiles.length) {
        successCallback();
        return;
    }
    var file = testFiles[index];
    var msg = 'create testFiles[' + index + ']: ' + file.path;
    var callback = function(entry) { setupFiles(root, index + 1, successCallback); };
    if (file.directory) {
        root.getDirectory(file.path, {create:true}, callback, errorCallback(msg));
        return;
    }

    root.getFile(file.path, {create:true}, function(entry) {
        entry.createWriter(function (writer) {
            writer.truncate(file.size);
            writer.onerror = errorCallback('truncate ' + entry.fullPath);
            writer.onwriteend = callback;
        }, errorCallback('create writer for ' + entry.fullPath));
    }, errorCallback(msg));
}

function startTest() {
    setupSourceFileSystem(function() {
        setupFiles(sourceTestDirectory, 0, function() {
            setupDestFileSystem(function() {
                log('Successfully setup test environment.');
                runTests(0, function() {
                    log('Successfully ran ' + tests.length + ' tests.');
                    cleanupDirectory(destTestDirectory, function() {
                        cleanupDirectory(sourceTestDirectory, function() {
                            log('Cleanup done.');
                            if (window.testRunner)
                                testRunner.notifyDone();
                        });
                    });
                });
            });
        });
    });
}

if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

startTest();

</script>
</body>
</html>
