third_party/WebKit/LayoutTests/fast/filesystem/cross-filesystem-op-expected.txt

This tests basic cross-filesystem operations.
Successfully setup test environment.
file__0:Temporary: reset /test
Testing copyTo apple.gif: /resources => /test
PASS: "apple.gif" == "apple.gif"
PASS: "1476" == "1476"
Testing copyTo mozilla.gif: /resources => /test
PASS: "mozilla.gif" == "mozilla.gif"
PASS: "2593" == "2593"
Testing copyTo drop-file-svg.svg: /resources => /test
PASS: "drop-file-svg.svg" == "drop-file-svg.svg"
PASS: "109" == "109"
Testing copyTo copy-backslash-euc.html: /resources => /test
PASS: "copy-backslash-euc.html" == "copy-backslash-euc.html"
PASS: "478" == "478"
Testing copyTo test_directory/test.txt: /resources => /test
PASS: "test.txt" == "test.txt"
PASS: "13" == "13"
file__0:Temporary: reset /test
Testing copyTo test_directory: /resources => /test
PASS: "test_directory" == "test_directory"
PASS: "1" == "1"
PASS: "test.txt" == "test.txt"
PASS: "13" == "13"
Successfully ran 2 tests.
Cleanup done.
